Add support for simple recipes, including get_tiddlers() and determine_bag() methods.
A recipe is a list of bags and an optional filter string for each bag.

get_tiddlers traverses the bags and filters generating a list of 
unique tiddlers that are the result of the recipe.

determine_bag() reverses the list and figures out the most likely
bag for a provided tiddlers to be saved, and returns it.

Ignoring persistence for the time being.

git-svn-id: http://svn.tiddlywiki.org/Trunk@3678 bb0f57cd-c710-0410-a8fb-e8bc1be0d679