reorient the filtering of tiddlers out of bags so that filtering is done by the filter module, rather than being a method on bag.
We want to keep the data classes quite simple, easy
to manipulate. 

It may make sense to pull some of the methods that do
filtering out of recipe and put them somewhere else,
but for now we'll call it okay because a recipe does
contain both a bag and a filter. The above change was
made because a bag does _not_ contain a filter, it is
filter_ed_.


git-svn-id: http://svn.tiddlywiki.org/Trunk@3742 bb0f57cd-c710-0410-a8fb-e8bc1be0d679