First pass at storing bags and tiddlers to disk. Incomplete. Tests not rigorous enough yet, recipes are not stored, and interface to methods not yet solid.
The store system copies the style of the serializer system
to create an eventually pluggable system of ways to store
things. The first implementation just stores into the file
system, serializing tiddler content using the text serializer.

This may appear to be noisy and a bit non-dry, but it means
that data objects are separated from their persistence model
making it easy to test things as well as easy to create 
new modes of storage.

Next step is to make the tests far more robust, including
proper exception throwing and handling.

Would like the save(bag, tiddlers) method to take either
one or a list of tiddlers, not only a list.

After this is a bit more solid, can add in recipe storage
(serializer recipe as text to somewhere on disk) and then
expolore getting and deserialization (storage to object
transformation).

Then we can see about building some proper controller code
that hooks it all together.


git-svn-id: http://svn.tiddlywiki.org/Trunk@3837 bb0f57cd-c710-0410-a8fb-e8bc1be0d679