Do some roundtrip testing: object into store, out of store.
Including using a stored recipe to figure out which bag
we need to store something in. This revealed that 
Recipe.determine_bag() wasn't really right. It's
algorithm would only return true if the tiddler of the
same name was already in the bag, and what we want is
if the tiddler with the same name were in the bag,
would it filter out properly. So that has been fixed.
I believe this is correct, but it is hard to be sure.

We are close enough now to creating web tools so
we'll do that next, after some filesystem cleanup.

git-svn-id: http://svn.tiddlywiki.org/Trunk@3860 bb0f57cd-c710-0410-a8fb-e8bc1be0d679