Capture HTTP exceptions in middleware and do the right thing.
Just as a start. This can be handled differently throughout
so there are lots of changes.

To make this go, had to actually make wsgi middleware, which I
hadn't been using before. So the core app is a wsgi app, and
then there are middleware wrappers. This should be useful
down the road.

Now need to catch up the rest of the code to use the stuff
properly.

git-svn-id: http://svn.tiddlywiki.org/Trunk@3988 bb0f57cd-c710-0410-a8fb-e8bc1be0d679