Add support for [sort[<field>]] and [count[<number>]] to the filtering spec. This was harder than it should have been.
The result is we can create URLs (and recipes) like this:

http://0.0.0.0:8080/bags/bag4/tiddlers?[tag[tagthree]]%20[tag[tagtwo]]%20[sort[-title]]%20[count[2]]

Or more interestingly, this:

http://0.0.0.0:8080/bags/ChrisContent/tiddlers?[tag[WorkBlog]]%20[sort[-created]]%20[count[10]]

Which would be something like the ten most recently created items tagged
WorkBlog. When I load such a page I could edit the content. When you load
such a page, you could only read it, but you _could_ write content that was
saved in the ChrisContentComment bag.

(and etc)

Must be time to add support for time to the tiddlers.


git-svn-id: http://svn.tiddlywiki.org/Trunk@4050 bb0f57cd-c710-0410-a8fb-e8bc1be0d679