When no bag or recipe is available, throw and error when trying to put a tiddler.
In general there is an incompleteness to the defensive
of the SyncMachine in the face of error conditions. The
adaptors try to announce when there are error conditions
but the SyncMachine doesn't realy do much with the information.

for example if a putTiddler results in a 404, the sync table
is still updated to show that the tiddler was updated on the
server.



git-svn-id: http://svn.tiddlywiki.org/Trunk@4227 bb0f57cd-c710-0410-a8fb-e8bc1be0d679