Some notes on how to handle revisions in NOTES and some wide ranging and very incomplete scribbles about users, authentication and authorization in AUTH.
The primary challenge with AUTH seems to be with dealing
with situations where XHR actions in Javascript need to
deal with unauth responses from the server in a way that
does _not_ cause or require the browser to be taken to
another page yet still allows the user to engage with and
select from a menu of delectable authentication possibilities.

Must confer with the ccTiddly crowd at some point.


git-svn-id: http://svn.tiddlywiki.org/Trunk@4604 bb0f57cd-c710-0410-a8fb-e8bc1be0d679