tiddlyweb - use wikklytext to add a rudimentary serverside tiddlywiki text to HTML transformation when requesting a tiddler as HTML. WikklyText's format is not quite identical to TiddklyWiki but is close enough to give a good sense of what's going on. The biggest difference is a lack of WikiWords, but we can probably figure that out (I've written to the author).
If the wikklytest libraries are present, the tiddler is formatted
to HTML, with links between tiddlers, and a link at the top of
the page to all the tiddlers in the recipe or bag (depending
on how the tiddler was addressed).

If the wikklytext libraries are not present, fall back to the 
old way of doing things.

Eventually this should allow a simple server side wiki, no
tiddlywiki required. But let's not jump the gun.


git-svn-id: http://svn.tiddlywiki.org/Trunk@5335 bb0f57cd-c710-0410-a8fb-e8bc1be0d679