tiddlyweb - extensive refactorings to start the process of adding a authentication challenge system. A new middleware has been created: PermissionsExceptor (in tiddlyweb.auth), and the stack of WSGI apps (in tiddlyweb.web.serve) has been reordered so it catches exceptions before the HTTPExceptor.
If a ForbiddenError is thrown, it is turned for the time
being into a HTTP403 exception. If a UserRequiredError is
thrown it will eventually present the requestor with some
kind of challenge. Right now it mumbles a bit of something
and blows up.

The PermissionsExceptor is intentionally separated from the
UserExtract middleware, so that both can operate independently.

Doing this work causes some circular dependencies, requiring
the creation of a tiddlyweb.web.util, which will be committed
in a moment.


git-svn-id: http://svn.tiddlywiki.org/Trunk@5564 bb0f57cd-c710-0410-a8fb-e8bc1be0d679