tiddlyweb - First start at having a User object. Does nothing so far other than get stored.
This model of the user is, for the time being, very simple.
The idea is that complicated users live elsewhere. All we
need is someway to make a relationship with that thing that
lives elsewhere. As we figure out what that means, will extend
the data model accordingly. For now all a user is:

* a usersign (thusly named to be more accurate and encompassing
  than username)
* a note
* a string pointing at an auth_system

This last will be used to do authentication and credentials
extraction. Or it may be split in two, as those aren't the same
thing.

So expect a lot of change in this handling.

git-svn-id: http://svn.tiddlywiki.org/Trunk@5637 bb0f57cd-c710-0410-a8fb-e8bc1be0d679