tiddlyweb - commit code and instructions for running TiddlyWeb under googleappengine.
This required some fairly extensive changes to 
tiddlyweb.web.server to accomodate setting server_host,
server_store and the middleware stack. The presence
of several global statements should highlight that it
is a big mess and needs some cleanup.

Note that this is the first example of a pluggable
Store for tiddlyweb, and it works in a nice straightforward
fashion. There are certainly ways to do this with less
code, but this adaptor style means that extension is 
easy.



git-svn-id: http://svn.tiddlywiki.org/Trunk@5667 bb0f57cd-c710-0410-a8fb-e8bc1be0d679