tiddlyweb - move various static variables into a config dict on tiddlyweb.web.serve. This gets stuck into the WSGI environ as tiddlyweb.config for use throughout the system by the 'Configurator' middleware.
The 'wiki' serialization made a direct call back to the config
to get server_host information, because it does not (yet)
have access to the environ. Now that the serializers are classes,
the environ could be passed to be an available attribute. I'll think
on this.

Eventually Configurator will be its own separate module, have 
a default config, and read a bona fide configuration file

This work is being done to make it easier to declare what store,
extensions and serializers an installation is supposed to support,
inspired by the GAE work.


git-svn-id: http://svn.tiddlywiki.org/Trunk@5680 bb0f57cd-c710-0410-a8fb-e8bc1be0d679