tiddlyweb - Add support for deleting tiddlers, with bag policy checking.
We've got some ambiguity about how permissions should be handled
for bags and recipes, so leaving them without a DELETE method. 
When the use cases bubble to the surface we'll figure it out.

This changed encouraged yet more cleanup in the dynamic method
discovery in tiddlyweb.store.

Now to get the GAE store deleting.


git-svn-id: http://svn.tiddlywiki.org/Trunk@5703 bb0f57cd-c710-0410-a8fb-e8bc1be0d679