tiddlyweb - exist a for loop as soon as we have the results we once.
I did some experimenting here to see if it was possible
to use generators when we are working with collections
of tiddlers. This proved to not work well with the filtering
architecture which suggests that the filtering architecture could
do with some changes. The filters are already way to unclear
so heading off into the world of generators and coroutines may
not be advisable but, you know. 


git-svn-id: http://svn.tiddlywiki.org/Trunk@5712 bb0f57cd-c710-0410-a8fb-e8bc1be0d679