Update the TODO with some information about searching.
Before we go too far down the searching road we need to have
some discussion about keeping the distinction between searching
and bag filtering clear.

The model described in the TODO assumes and requires that all the 
smarts for the searching is in the store, the rest of TiddlyWeb is
completely ignorant. This is how it should be.



git-svn-id: http://svn.tiddlywiki.org/Trunk@5768 bb0f57cd-c710-0410-a8fb-e8bc1be0d679