tiddlyweb - create a TiddlyWebExplorerPlugin which provides a <<tw_explore>> macro. Place that in a tiddler and the current server (in the location bar) will be queried for a list of recipes. Each of these will be be added as <li> on <ul>. Then for each recipe, query the server for the recipe info, display that, then display, as links to the server, all the tiddlers produced by the recipe. Use <<tw_explore $server_url>> to get the information from an arbitrary server (only works from a file based tiddlywiki).
zebag is a handy bag to symlink into the store of a running 
TiddlyWeb server somewhere else, for the sake of testing.

The point of this plugin is to demonstrate ways in which the 
TiddlyWebToolsPlugin can be used to inspect and use the data
found on a TiddlyWebServer, without necessarily incorporating
that information into the running TiddlyWiki.

git-svn-id: http://svn.tiddlywiki.org/Trunk@6200 bb0f57cd-c710-0410-a8fb-e8bc1be0d679