tiddlyweb - resolve http://trac.tiddlywiki.org/ticket/731
Provide a download capability in the URL when sending a list of tiddlers.

At the moment this only works for the wiki serialization. As I write
this I realize there's not particular reason why this should be so. So
I'll change that on the next commit.

To get it go, form a URL as:

http://0.0.0.0:8080/recipes/AutoTiddlyWeb/tiddlers.wiki?download=foo.html

The argument to download is the filename the content disposition header
will suggest to the browser. It's probably best if it has an extension that
is a) meaningful to the browser b) matches the content type of the thing
being downloaded.


git-svn-id: http://svn.tiddlywiki.org/Trunk@6384 bb0f57cd-c710-0410-a8fb-e8bc1be0d679