tiddlyweb - Introduce the concept of the HTMLPresenter
There's a running tension between TiddlyWeb's design
as a pure web service for storing Tiddlers for use
in the dynamic construction of TiddlyWiki (and TiddlyWiki
like tools) and as a minor web application for presenting
Tiddlers and information about them in a somewhat useful
fashion and style (say as rendered HTML).

For example, if the service creates HTML which is lovely
for people, it is often the case that is less lovely for 
tools that want to dork with it. Tools might like fragments
of HTML documents, while people using browsers might like
full HTML documents with style, and maybe even some furniture
around the edges to make the representations contained 
therein contextualized and useful.

In the early days of TiddlyWeb the HTML serialization got
a bit crufty because it took on the job of creating human
usable content. This was necessary to get things rolling
but is a larger job than the serializations were built to do.
The _sole_ outgoing job of the serializtaion is to take a resource
and turn it into some representation: HTML, text, cake.

It is not the job of a serialization to make the representation
pretty and presentable.

Thus the HTMLPresenter. The HTMLPresenter is a piece of WSGI
middleware that runs after the main TiddlyWeb app, and if the
conditions are right, wraps HTML content with friendly 
presentableness, such as <body> tags, or maybe some CSS.

The html serialization puts tiddlyweb.title and tiddlyweb.links
into the environment for the HTMLPresenter to use if it wants.

Because HTMLPresenter is part of the WSGI stack in TiddlyWeb
and because that stack is entirely configurable the HTMLPresenter
can be replaced with something else that does more or different
work than the current HTMLPresenter. One option would be a thing
that uses a templating engine. Because the presenter has access
to the wsgi environment, it can make choices based on the current
user, the URL of the request, etc etc to create a fully dynamic
presentation if so desired.

It is also possible to remove the HTMLPresenter from the stack
entirely preserving the raw HTML and making it possible for
some other web server entirely to handle the presentation.
For example some fully framework content management system
that uses TiddlyWeb for some aspect of its content store.

This is a good thing.

The introduction of this features leads to a bit of confusion
over where to adjust the HTML generated by TiddlyWeb. Before
the only option was to replace or subclass the HTML serialization.
Now there are two options, the serialization and some form
of HTMLPresenter. So how do you choose which to use?

Basically if you want to change the HTML representation of a 
resource, you change the serialization. If you want to change the
HTML that surrounds, couches, cuddles an HTML representation,
do that in the HTMLPresenter.

Improvements to HTMLPresenter will come.

Other changes here include some adjustments to how the 
wsgi apps are called to make sure exceptions are properly
handled, reported and cascaded up the stack. We need to
use exc_info correctly and weren't.



git-svn-id: http://svn.tiddlywiki.org/Trunk@6564 bb0f57cd-c710-0410-a8fb-e8bc1be0d679