tiddlyweb - some intial pokings at providing support the notion of user roles. Current implementation is incomplete as there is no support in bag policies for thinking about roles.
This is very much an "enough rope to hang yourself" kind
of implementation. Each user gets a set associated with it
which lists the roles that user has. There is no canonical
list of roles, so no effort is made to maintain referential
integrity: that's the job of the administrator of the system.

So if you assign cdent role ADMIN but fnd role admin, they
are in different roles and nothing is going to tell you
otherwise (except for things breaking later).

Next: determine how to utilize this information with polices
and elsewhere. Probably by changing policy.allows to take
the environment instead of a usersign, which I don't like
as it is not constrained. Will try to think of something else.


git-svn-id: http://svn.tiddlywiki.org/Trunk@6577 bb0f57cd-c710-0410-a8fb-e8bc1be0d679