tiddlyweb - inject SiteTitle and SiteSubtitle into <title /> of the outgoing wiki if one or both is present, or if the wiki is made up of just one tiddler, use that tiddler's title, or use 'TiddlyWeb Loading'.
We use the HTML serialization and BeautifulSoup to do the
HTMLification and subsequent plain textification of the
wikitext in SiteTitle and SiteSubtitle. If that stuff fails
for some reason we use the straight wikitext.

Because we are calling the HTML serializer, the environment
gets messed with, so we need to unmess it when we get
back to the wiki serializer or suffer the consequences. This
points out some of the problems with the way the HTMLPresenter
works and it also suggests that the HTMLification used by the
HTML serializer (and now by the wiki serializer too) should be 
in its own module. Which is somehow funny or ironic. Not sure
which. I'm going to ask Winona Ryder which one it is.

This change feels somehow messy. I'm not entirely sure why.

It should, however, make the MarkupPreHead stuff easier, now
that the barrier has been broken.


git-svn-id: http://svn.tiddlywiki.org/Trunk@6605 bb0f57cd-c710-0410-a8fb-e8bc1be0d679