tiddlyweb - fix some twoter bugs noted by users:
twotes from the same page clobber one another, now they 
don't.

Pages with titles with | in them were messing up the server
side wikifier. Now they don't.

The fix for the first one required an interesting hoop jump:
The easiest way to tell if a tiddler already exists in the
store is to get() it. If a NoTiddlerError exception is raised,
its not in there and we can deal accordingly. However, if it
is there, then now the tiddler object you just get()ed is now
populated with the data from the store, not the data you
got from the client. Ooopth. So we have to get a tiddler of the
same name and bag but a different object.

git-svn-id: http://svn.tiddlywiki.org/Trunk@6925 bb0f57cd-c710-0410-a8fb-e8bc1be0d679