tiddlyweb - from_svn, a utility for pulling plugins, tiddlers or the plugins and tiddlers represented by a cook recipe into a bag in TiddlyWeb. The script is given the name of a bag and one or more URLs--*.tiddler, *.js or *.recipe--and does the right thing to retrieve and massage the content such that it can be injected into the TiddlyWeb store. In the case of a recipe, it will recurse multiple recipes.
Note that while the code is dependent on the format of the 
stuff stored in subversion (i.e. it must know about it) it
is not dependent on the format of the store, it just calls 
store and the TiddlyWeb library code handles getting things
to disk, or google, or wherever.

git-svn-id: http://svn.tiddlywiki.org/Trunk@6960 bb0f57cd-c710-0410-a8fb-e8bc1be0d679