tiddlyweb - Add a policy to recipes to do access control. We really only use manage and read at the moment, but there's room for expansion.
to manage means "to alter the policy of a new or existing
recipe" (or bag as well). This means you need manage to
write a recipe or bag, or delete one. read allows you to
see what's there.

Some tweaking is probably going to be needed here.

Policy was extracted to its own module and reused as it
already provided the user<->constraint mappings and 
exception raising.

Next: bag_createp() and recipe_createp() handling.


git-svn-id: http://svn.tiddlywiki.org/Trunk@7066 bb0f57cd-c710-0410-a8fb-e8bc1be0d679