tiddlyweb - A comment from FND inspired me to make a simple plugin system for TiddlyWeb. This is done in an explicit way:
Config has a system_plugins key. It's value is a list of 
strings representing modules in python's sys.path. For
each plugin, init() is called in the module, with the
config dictionary passed as an argument. Within init()
whatever needs to happen can be happen.

The init() methods are called when load_app() is called, 
which is when the web service is started. For mod_python,
cherrypy, googleappengine this is once, when things are turned
on. For CGI this is with each request. That is it happens
at the start of a new process.

A simple tiddlyweb.debugplugin is included for demonstration
purposes.

The reason for having a thing like this is so we can easily
modify existing code while still using the normal startup
methods. All we have to do is make an entry in our tiddlywebconfig.py
file.

git-svn-id: http://svn.tiddlywiki.org/Trunk@7159 bb0f57cd-c710-0410-a8fb-e8bc1be0d679