tiddlyweb - throwing a spike into the notion of storage of binary stuff inside a tiddler. If we don't know the content-type of the incoming thing (if we can't serialize it) we just store it as base64 encoded string.
This works in a very raw way but only from the web
interface, so we need some tweaks.

Also, if you make a tiddlywiki, you end up with a tiddler
with a bunch of base64 stuff in it. We can make some plugins
go with this to try data uris. 

We shall see. More to come.



git-svn-id: http://svn.tiddlywiki.org/Trunk@7278 bb0f57cd-c710-0410-a8fb-e8bc1be0d679