tiddlyweb - the start of a test that demonstrates the HTTP API in a more clear way than the existing tests.
you can run this as either a normal py.test test:

  py.test test/test_http_api.py

and it will run some assertions abased on the TESTS
list in the test. Or you can run it as its own script:

  python test/test_http_api.py

and it will produce a list of the test names, urls and
brief descriptions, which might be a nice learning
tool.

git-svn-id: http://svn.tiddlywiki.org/Trunk@7635 bb0f57cd-c710-0410-a8fb-e8bc1be0d679