tiddlyweb - twome, a collection of plugins and other tools which may be useful in creating a tiddlyweb based web site. The idea is to use these things to create a tiddlyweb-based site to host TiddlyWeb's advertising of itself.
Checking this in to show where I've got to. It's the sort of
stuff that will want for some user feedback and more thought
about experience goals.

admin.py is a plugin that provides a framework for a superuser
to list all bags, recipes, users etc and do things to them.
Much of that work is better served on the client side, I think,
so I've left it basically empty, so we have the structure to work
from.

faqtwer is supposed to be a way to use a single bag as a source
of faq questions and answers, using tags as a category system.
Again, ui left lame for more excitement later. These things may
be better served presented as a tiddlywiki.

homer is a plugin that provides a home page, replace the TiddlyWeb
default. Basically just a road to a template.

plugins is a set of decorators that were useful in the other plugins.
This is yet another fun python trick that abstracts away duplicated
code so you don't have to think about it.

spowt2 is a slight refactoring of the original spowt to use the
new decorators

static is here because it was needed. I probably shouldn't have checked
it in.

I hope this gives some idea of how plugins can be lumped together
for interesting results.

git-svn-id: http://svn.tiddlywiki.org/Trunk@7938 bb0f57cd-c710-0410-a8fb-e8bc1be0d679