tiddlyweb - update tests and wiki serialization to reflect some formatting changes in tiddlywiki empty.html file.
This is exactly why some of these tests exists, but it also
points out a problem with the way the title stuff is handled.
Should probably change to a dom processing kind of way. Using
a templatized empty.html is less desirable because then we
can't use arbitrary TiddlyWikis as the source.

git-svn-id: http://svn.tiddlywiki.org/Trunk@8327 bb0f57cd-c710-0410-a8fb-e8bc1be0d679