Completely re-architected internals of comments plugin for better TiddlyGuv/CCTiddly related permissioning model.
Previously, adding a new comment meant changing its precursor (ie the daddy pointing to its first child, or the previous tiddler pointing to its next sibling). You can see how this could cause permissioning problems - a commenter often doesn't have permission to save its precursor!

The new model means a new comment only has to save itself. It effectively reverses the situation so the comment is a kind of view on the rest of the model. In practice, this means the comment points to its precursor. For efficiency, all comments now contain a "daddy" field, which is filled (regardless of whether this is the first child comment) and a "root" field. In addition, all but the eldest child has a "prev" field pointing to its previous sibling.

For efficiency purposes, a "treeize" method is sometimes used, which sets "next" and "firstChild" values on all the tiddlers in the hierarchy. This slightly goes against the TiddlyWiki design principle of making everything explicit through field names. This is necessary because if only the "prev" and "daddy" fields were used, algorithm time would be O(N^2) or worse, where N is the entirety of tiddlers in the store, not just the comments in the hierarchy. So, while being wary of the evils of premature optimisation, I decided in this case to err on the side of efficiency.

IMPORTANTLY, the "treeize" method is only used to form a read-only view of the hierarchy, and only at the start of each function that needs it, i.e. it should be transparent externally (other than the fact that tiddler memory objects do have "next" and "firstChild" pointers attached to them, which may well be out of whack - we might consider explicitly removing those for pristine transparency purposes). The key point and justification here though is the *ready-only* nature of this in-memory data structure. There is no attempt to change the tree and then resync all the fields (unless the first incarnation of this plugin). It is simply there to make traversal operations both performant and elegantly coded.



git-svn-id: http://svn.tiddlywiki.org/Trunk@8345 bb0f57cd-c710-0410-a8fb-e8bc1be0d679