tiddlyweb - tune up the gzipper a bit:
make it a plugin that auto installs itself
in the response_filter stack, rather than
the user needing to adjust the config.

log when it is choosing to compress

have it only compress when the response status is 200.

There appears to be some breakage going on with gzipper
when trying to use the openid challenger, so it seems
it is not happy dealing with redirects, so this is not
yet a stable thing.

git-svn-id: http://svn.tiddlywiki.org/Trunk@8671 bb0f57cd-c710-0410-a8fb-e8bc1be0d679