tidldyweb - enable some a11y handling in the case where the wiki generated by TiddlyWeb is delivered to a UserAgent that does not support javascript: put a link to the HTML representation of the list of tiddlers generated by the recipe or bag.
This is potentially a bit wonky as sometime the wiki generated
is not from a recipe or bag but is generated by a plugin, but
we'll deal with that bug when it becomes one.

This should make tiddlyweb content a bit more indexable.

Including it now as an experimental feature that may be deprecated
in the future.

git-svn-id: http://svn.tiddlywiki.org/Trunk@8850 bb0f57cd-c710-0410-a8fb-e8bc1be0d679