tiddlyweb - add support for filter on extended fields. The syntax follows from the other fields:
   [status[hot]]

will get those tiddlers that have an extended field 'status' with
the exact value 'hot'.

   [!status[hot]]

will get those that do not have that field or do have that field but
not a value of 'hot'.

The implementation is the simplest possible: first we look in
FILTER_MAP. If we get a KeyError, then we try to find a matching
field.

git-svn-id: http://svn.tiddlywiki.org/Trunk@9057 bb0f57cd-c710-0410-a8fb-e8bc1be0d679