tiddlyweb - make twanager properly report a non-zero exit code when it fails to do what you asked it to do. This makes it better for scripting.
In the process also make sure all prints that might happen while
twanager gets called go to sys.stderr instead of stdout. Again to
make it more compatible with scripting.

Removed a test in test_manage because now calling that code calls
sys.exit(1), which exits the test harness, which we don't want.

git-svn-id: http://svn.tiddlywiki.org/Trunk@9428 bb0f57cd-c710-0410-a8fb-e8bc1be0d679