Attempt at a client-side copy function to copy a tiddler from one bag to another, handling conflicts that arise, by michael mahemoff, jon lister, ben gillies.
It works by changing server workspace to the destination, saving the tiddler (using the adaptor), and reverting back the server workspace and revision ID. This "happy path" scenario involves no preliminary GET call to inspect the state of the tiddler, if it exists, in the destination bag.

In the "unhappy" case where the tiddler does exist in the destination bag and has a different revision ID - triggering a conflict - this plugin will simply force a replacement by retrieving and inspecting the destination tiddler, and updating the source tiddler's temporary revision ID. It's envisioned that future versions will accept a "force" parameter, and if "not force", the user will be able to see a diff and decide how to handle the conflict.



git-svn-id: http://svn.tiddlywiki.org/Trunk@9597 bb0f57cd-c710-0410-a8fb-e8bc1be0d679