unaforms - get the forms handling working for unauthenticated users
When of the primary use cases is to be able to give a URL
to joe random and for them to then be able to work with the
form data. The changes in this commit are step in that
direction, with more steps to come to support further authentication
models. These changes set the stage for all the change by changing
the fundamental model for how things work which should prove more
flexible.

In the original model there was a form id for recipe, and then
when an authenticated user visited a form they got their own special
bag created for them, and their wiki was configured to save content
to that bag.

In the new model we get rid of the form id for a recipe. And instead
use the recipe name (prepended by 'form.') to identify a form. Then
we create URLs for accessing those forms, as needed, from the admin
interface, the URLs identify the form name involved and a bag. The
URL is likely unguessable because it is a UUID, but includes enough
information to associate it with the particular recipe that makes up
a form. The generate wiki is configured to safe to a bag based on the
UUID. So an admin has a few choices:

* create a new UUID based URL for every person they want to fill out
  the form and mail them that URL
* create only some UUID based URLs and send the relevant one to a
  group of people who should be able to see each other's data

This change makes a useful separation between the recipe that generates
the form (the form template) and accessing it by some identifier
which may or may not require authentication. The current implementation
does not require authentication, but there is room to shoehorn that in.
In the previous implementation it was required. Now it is not.

With this change we've gotten rid of the yaml.db and the tools for
manipulating it. All that is needed to run the system is:

* a user with the admin role who can see /forms from which they
  generate the UUID based URLs
* a recipe that has the various bits which make up the form-template

This latter part requires smoother automation. That will come.

git-svn-id: http://svn.tiddlywiki.org/Trunk@9738 bb0f57cd-c710-0410-a8fb-e8bc1be0d679