Created Twispa (Twitter SPA) - proof-of-concept SPA to post from multiple twitter accounts. (or should it be called "Twitspam" ;)
This builds on the spa demo app to show a real example of the twFile plugin in action

The app is working in Firefox - lets you add several twitter usernames and passwords, gives you a dropdown and tweet prompt and posts to Twitter. Thanks to FND for showing me there was a much simpler path for twitter authentication than faking a login submission! (ie basic auth)


git-svn-id: http://svn.tiddlywiki.org/Trunk@10329 bb0f57cd-c710-0410-a8fb-e8bc1be0d679