TiddlyWebAdaptor: further tweaked ETag handling for new tiddlers
It has now been agreed that new tiddlers can have an ETag to indicate
that edit conflicts should be detected.

This ETag takes the following form:
    <bag>/<tiddler>/0

Note that PUTs to a recipe lack an ETag, thus leaving a possibility for
accidental clobbering.
This should be addressed at the application level - ideally by PUTing
directly to bags instead of recipes.

git-svn-id: http://svn.tiddlywiki.org/Trunk@10391 bb0f57cd-c710-0410-a8fb-e8bc1be0d679