InfoBoxPlugin - a new plugin to provide wikimedia-like infoBoxes.
In the simplest case, user just writes <<infoBox>>this is my message>>. More sophisticated use comes with a definition tiddler; <<infoBox tip>>think first>> will invoke the "tipInfoBox" tiddler, which contains slices determining the box's appearance. One slice possibility is "message", which means the infoBox will have a fixed message, which means the macro's "inner text" is ignored, which means you can just invoke the macro as <<infoBox warning>>>> .

(Note the use of a single closing >> with no matching << is a little, shall we say, unconventional...but is used for consistency with the long-existing <<gradient>>my message here>> macro. And short of diving head-first into the wikification code, it's the only way I know how to an "inner" text block.



git-svn-id: http://svn.tiddlywiki.org/Trunk@10558 bb0f57cd-c710-0410-a8fb-e8bc1be0d679