Comments Plugin - Added important fix to ensure comments can be submitted in parallel, i.e. user A opens wiki, user B opens wiki, user A or B submits comment, other user submits comment
To be as robust and scaleable as possible, there's no check upon submission (which would still run into fine-grained race conditions, in the absence of a specialised server plugin). The only difference is the treeify algorithm, used to display the new tiddlers. It allows for the possibility of 2+ tiddlers pointing back to the daddy (implying each tiddler thinks daddy has only one child, itself), or 2+ tiddlers pointing back to the same "2nd last sibling" (implying each tiddler thinks its the last sibling).

At present, the second and latter tiddlers in this confused state will fall all the way to the end of the line. I want to put in a fix to use timestamps so the confused tiddlers will appear in order of time they were saved.


git-svn-id: http://svn.tiddlywiki.org/Trunk@10953 bb0f57cd-c710-0410-a8fb-e8bc1be0d679