TiddlyWebAdaptor: ensured getTiddler sets all standard fields
Custom fields server.bag and server.recipe fields were not being set.
While this might have been intentional at the time, relying on
workspace, it now seems like an omission and inconsistency.

For example, in the context of TiddlySpace, a tiddler's bag is
significant, so the workspace abstraction alone is insufficient.

also added creator support and simplified workspace calculation

git-svn-id: http://svn.tiddlywiki.org/Trunk@12075 bb0f57cd-c710-0410-a8fb-e8bc1be0d679