TiddlyWebAdaptor: added flag to suppress ETag generation
if server.page.revision is "false", no ETag will be generated

While somewhat hacky, this is required for putTiddler to allow forcing
the creation of a new revision by omitting If-Match headers.

Prompted by a "publish" command which needs to create a tiddler revision
irrespective of whether any previous revision exists.

git-svn-id: http://svn.tiddlywiki.org/Trunk@12105 bb0f57cd-c710-0410-a8fb-e8bc1be0d679