BinaryTiddlersPlugin: disabled auto-linking of WikiWords
The primary purpose here is to exclude binary tiddlers from the list of
missing tiddlers. However, the TiddlyWiki core currently does not
provide sufficient encapsulation - ideally, the (badly named) changed
Tiddler method and/or the getMissingLink TiddlyWiki method would provide
a hook - so resorting to autoLinkWikiWords. Consequently, bracketed links
in binary tiddlers will *not* be suppressed.

git-svn-id: http://svn.tiddlywiki.org/Trunk@12334 bb0f57cd-c710-0410-a8fb-e8bc1be0d679