TiddlyWebAdaptor: removed post-PUT GET request unless required
The (only?) reason putTiddlerCallback triggered a GET was to update the
revision number, which in turn was required to generate the ETag for
subsequent PUTs or DELETEs.

Legacy browsers aside, the ETag can be retrieved directly from the PUT
response, so it is now cached and reused if available.

NB: Once TiddlyWebWiki's serialization includes the server.etag field,
generateETag should not be required anymore, reducing the tight coupling
by not requiring the client to be aware of the ETag's structure.

git-svn-id: http://svn.tiddlywiki.org/Trunk@12363 bb0f57cd-c710-0410-a8fb-e8bc1be0d679